requirements/base.txt
Django==3.2.5
django-filter==2.4.0
djangorestframework==3.12.0
drf-yasg==1.20.0
gunicorn==20.1.0
Pillow==8.3.1
